requirements.txt
asgiref==3.9.2
Django==5.2.6
django-stubs==5.2.5
django-stubs-ext==5.2.5
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
gunicorn==23.0.0
mysqlclient==2.2.7
packaging==25.0
pillow==11.3.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
sqlparse==0.5.3
types-PyYAML==6.0.12.20250915
typing_extensions==4.15.0
